+++ daemon/assets/ghc-specter.css
@@ -81,7 +81,7 @@ nav {
     width: 50%;
 }
 
-.column.is-one-fifths {
+.column.is-one-fifth {
     width: 20%;
 }
